fix(contract): distanceToCentreDays is a midpoint distance, not a day count
Real data: 10 of 17 texts overlapping one real photo carry 6.5, not the
integer this field was constrained to. It measures the distance between
two interval MIDPOINTS, not a count of days — an even-length span's
centre legitimately falls between two calendar days, and the backend
does not round it. This app's own domain/interval.ts#centreDistanceDays
rounds for its own use as a client-side sort tiebreaker; that rounding
was never a claim about what the server sends.

+++ src/api/contract/overlap.ts
@@ -21,7 +21,15 @@ export const OverlapInfoSchema = z.strictObject({
   textSpanDays: z.number().int(),
   /** Default sort: this sum, ascending. */
   totalSpanDays: z.number().int(),
-  distanceToCentreDays: z.number().int(),
+  /**
+   * A midpoint distance, NOT a day count — `.5` is a real, common value
+   * (the backend does not round it; confirmed against real data: an even
+   * span's centre falls exactly between two days). This app's OWN
+   * `domain/interval.ts#centreDistanceDays` rounds for its own use as a
+   * client-side sort tiebreaker — that rounding is local to that function,
+   * never a claim about what the server sends.
+   */
+  distanceToCentreDays: z.number(),
 });
 export type OverlapInfo = z.infer<typeof OverlapInfoSchema>;
